Recommend ADD --chown 1001:0 syntax if possible
Unfortunately, --chmod option was added into docker in v17, so it does not work
with docker we have in RHEL-7.

For supporting docker with our example Dockerfiles, we can use the older way,
but in README.md, where we recommend using podman (that supports the --chown option),
we can already use the newer shorter syntax that also does not need changing
USER to 0 that might be prohibitted.