This is a hack for moby engine.
From unknown reason, moby does not like when using glob in a directory
behind a symlink. In this case, 1.22-minimal is a symlink to 1.22 example
app and when using this example app, moby engine does not match *.html
files despite being there (confirmed that replacing glob with file names
works). Here we also expect that symlinks are only the minimal variants,
so therefore we remove -minimal suffix in case when the example app
directory is a symlink.